Refactor command_executor_main function and add execute_command function
This commit refactors the `command_executor_main` function in the `__init__.py` file and adds a new `execute_command` function. The `command_executor_main` function now uses the `get_command_specs` function to generate command specifications and dynamically create argument parsers. The `execute_command` function is responsible for executing the specified command with the given parameters. This refactoring improves code organization and makes it easier to add new commands in the future.